% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_atco_codes.R
\name{get_atco_codes}
\alias{get_atco_codes}
\title{Return a vector of all permitted ATCO region codes in NaPTAN}
\usage{
get_atco_codes()
}
\description{
Return a vector of all permitted ATCO region codes in NaPTAN
}
